pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = "A sample starter project using FastAPI and GraphQL"
authors = ["Eidan J. Rosado"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = "^0.13.4"
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
gql = "~3.0.0.a4"
fastapi-utils = "^0.2.1"
strawberry-graphql = "^0.93.4"
dataclasses-json = "^0.5.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
